# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for ATS-Tex Lab, a custom CV builder and ATS simulator.",
        "title": "ATS-Tex Lab API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/interview-prep": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate AI-driven interview questions based on CV profile and job description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interview Prep"
                ],
                "summary": "Create Interview Prep",
                "parameters": [
                    {
                        "description": "Interview prep parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CreateInterviewPrepRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_semmidev_atstex-lab_internal_domain.InterviewPrep"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            }
        },
        "/api/interview-prep/critique": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Provides real-time AI critique of a candidate's answer to a specific interview question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interview Prep"
                ],
                "summary": "Critique Interview Answer",
                "parameters": [
                    {
                        "description": "Candidate answer",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CritiqueInterviewAnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            }
        },
        "/api/interview-preps": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a history of interview preps run by the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interview Prep"
                ],
                "summary": "List Interview Preps",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_semmidev_atstex-lab_internal_domain.InterviewPrep"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/problem.Problem"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_semmidev_atstex-lab_internal_domain.InterviewPrep": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "jobDescription": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "profileId": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "tokensUsed": {
                    "type": "integer"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.CreateInterviewPrepRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "jobDescription": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "profileId": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.CritiqueInterviewAnswerRequest": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
...
        "problem.Problem": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "Detail is a human-readable explanation specific to this occurrence of the problem.",
                    "type": "string"
                },
                "instance": {
                    "description": "Instance is a URI reference that identifies the specific occurrence of the problem.",
                    "type": "string"
                },
                "status": {
                    "description": "Status is the HTTP status code for this occurrence of the problem.",
                    "type": "integer"
                },
                "title": {
                    "description": "Title is a short, human-readable summary of the problem type.",
                    "type": "string"
                },
                "type": {
                    "description": "Type is a URI reference that identifies the problem type.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}